requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
click==8.1.7
    # via uvicorn
fastapi==0.112.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
pyright==1.1.374
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.30.5
    # via -r requirements.in
pytest==8.3.2
black==24.8.0
python-dotenv==1.0.1
redis==5.0.8
pymysql==1.1.1
cryptography==43.0.0